+++ checkers/cwe/tests/suite/IntegerOverflowChecker/loop_iter_underflow_while.cpp
@@ -1,7 +1,7 @@
 
 void foo()
 {
-    int arr[10];
+    unsigned long long arr[10];
     unsigned long long i = 9;
     while (i > 0) {
         arr[i] = i;
@@ -11,7 +11,7 @@ void foo()
 
 void foo2()
 {
-    int arr[10];
+    unsigned long long arr[10];
     unsigned long long i = 9;
     while (i > 1) {
         arr[i] = i;
